Improve user feedback on operation.
Before it was possible to click the modal buttons
multiple time to trigger the same action again and
again. Also closing the modal did lose the indication
that there is still a operation running. Therefore, we
show the loading state on the action button and
disable it also until the operation is fulfilled.

Committed-by: Rene Pfeuffer <rene.pfeuffer@cloudogu.com>